backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
redis>=5.2.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.36
alembic>=1.14.0
bcrypt>=4.2.0
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9
email-validator>=2.0.0
psycopg[binary]>=3.2.0
Pillow>=10.0.0
reportlab>=4.2.0
